+++ src/RcsbUtils/AnnotationGenerators/AlignmentVariation.ts
@@ -1,5 +1,5 @@
 import {ExternalTrackBuilderInterface} from "../../RcsbCollectTools/FeatureTools/ExternalTrackBuilderInterface";
-import {Logo} from "../../RcsbFvWeb/RcsbFvGroup/Logo";
+import {Logo} from "./Logo";
 import {RcsbFvDisplayTypes, RcsbFvRowConfigInterface} from "@rcsb/rcsb-saguaro";
 import {SequenceCollectorDataInterface} from "../../RcsbCollectTools/SequenceCollector/SequenceCollector";
 import {RcsbAnnotationConstants} from "../../RcsbAnnotationConfig/RcsbAnnotationConstants";