+++ src/RcsbUtils/Translators/MultipleEntityInstanceTranslate.ts
@@ -1,7 +1,7 @@
 import {PolymerEntityInstanceTranslate} from "./PolymerEntityInstanceTranslate";
 import {PolymerEntityInstanceInterface} from "../../RcsbCollectTools/DataCollectors/PolymerEntityInstancesCollector";
 import {TagDelimiter} from "../TagDelimiter";
-import {Operator} from "../../Helpers/Operator";
+import {Operator} from "../Helpers/Operator";
 
 export class MultipleEntityInstanceTranslate {